+++ sam/metrics/__init__.py
@@ -1,17 +1,18 @@
-from .incident_recall import incident_recall, make_incident_recall_scorer,\
-    precision_incident_recall_curve
-from .mase import mean_absolute_scaled_error
-from .keras_metrics import keras_tilted_loss, keras_joint_mse_tilted_loss, \
-    keras_rmse, get_keras_forecasting_metrics, keras_joint_mae_tilted_loss
-from .train_mean_r2 import train_mean_r2
-from .tilted_loss import tilted_loss
+from . import keras_metrics, mase
 from .custom_callbacks import R2Evaluation
-from .quantile_evaluation import compute_quantile_ratios, compute_quantile_crossings
-
-from . import mase
-from . import keras_metrics
+from .incident_recall import (incident_recall, make_incident_recall_scorer,
+                              precision_incident_recall_curve)
+from .keras_metrics import (get_keras_forecasting_metrics,
+                            keras_joint_mae_tilted_loss,
+                            keras_joint_mse_tilted_loss, keras_rmse,
+                            keras_tilted_loss)
+from .mase import mean_absolute_scaled_error
+from .quantile_evaluation import (compute_quantile_crossings,
+                                  compute_quantile_ratios)
+from .r2_calculation import train_mean_r2, train_r2
+from .tilted_loss import tilted_loss
 
 __all__ = ["incident_recall", "make_incident_recall_scorer", "precision_incident_recall_curve",
-           "mean_absolute_scaled_error", "tilted_loss",
-           "keras_tilted_loss", "keras_rmse", "get_keras_forecasting_metrics", "train_mean_r2",
+           "mean_absolute_scaled_error", "tilted_loss", "keras_tilted_loss", "keras_rmse",
+           "get_keras_forecasting_metrics", "train_mean_r2", "train_r2",
            "compute_quantile_ratios", "compute_quantile_crossings"]